requirements.txt
asgiref==3.9.1
Django==5.2.5
django-cors-headers==4.7.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
Markdown==3.8.2
PyJWT==2.10.1
sqlparse==0.5.3
